# RSpec/SubjectStub. Refactor and decrease complexity