`RSpec/SubjectStub` takes about 35% of all the time spent in cops if run Rubocop on Rubocop's  own source code. 

PR addresses the performance problem mentioned in the https://github.com/rubocop-hq/rubocop/issues/8022 issue.

### Profiler reports

Profiler report (`stackprof`) before optimizaition:

```
stackprof tmp/stackprof-cpu-rubocop.master.with-config.dump --method 'RuboCop::Cop::Commissioner#trigger_responding_cops'

RuboCop::Cop::Commissioner#trigger_responding_cops (/Users/andrykonchin/projects/rubocop/lib/rubocop/cop/commissioner.rb:50)
  samples:   756 self (2.8%)  /   17379 total (65.3%)
  callers:
    17379  (  100.0%)  block (2 levels) in <class:Commissioner>
    17332  (   99.7%)  RuboCop::Cop::Commissioner#trigger_responding_cops
    16630  (   95.7%)  RuboCop::Cop::Commissioner#with_cop_error_handling
  callees (16623 total):
    17332  (  104.3%)  RuboCop::Cop::Commissioner#trigger_responding_cops
    16652  (  100.2%)  RuboCop::Cop::Commissioner#with_cop_error_handling
    5934  (   35.7%)  RuboCop::Cop::RSpec::SubjectStub#on_block
    1998  (   12.0%)  RuboCop::Cop::Layout::SpaceAroundMethodCallOperator#on_send
     374  (    2.2%)  RuboCop::Cop::Layout::SpaceInsideArrayLiteralBrackets#on_array
...
```

Profiler report (`stackprof`) afeter optimizaition:

```
stackprof tmp/stackprof-cpu-rubocop.master.with-config.rspec-optimized.dump --method 'RuboCop::Cop::Commissioner#trigger_responding_cops'

RuboCop::Cop::Commissioner#trigger_responding_cops (/Users/andrykonchin/projects/rubocop/lib/rubocop/cop/commissioner.rb:50)
  samples:   851 self (3.6%)  /   13455 total (56.9%)
  callers:
    13455  (  100.0%)  block (2 levels) in <class:Commissioner>
    13398  (   99.6%)  RuboCop::Cop::Commissioner#trigger_responding_cops
    12598  (   93.6%)  RuboCop::Cop::Commissioner#with_cop_error_handling
  callees (12604 total):
    13398  (  106.3%)  RuboCop::Cop::Commissioner#trigger_responding_cops
    12626  (  100.2%)  RuboCop::Cop::Commissioner#with_cop_error_handling
    2143  (   17.0%)  RuboCop::Cop::Layout::SpaceAroundMethodCallOperator#on_send
     395  (    3.1%)  RuboCop::Cop::Layout::SpaceInsideArrayLiteralBrackets#on_array
     305  (    2.4%)  RuboCop::Cop::Layout::SpaceBeforeFirstArg#on_send
     279  (    2.2%)  RuboCop::Cop::Style::RedundantSelf#on_block
     266  (    2.1%)  RuboCop::RSpec::TopLevelDescribe#on_send
     226  (    1.8%)  RuboCop::Cop::Layout::SpaceInsideReferenceBrackets#on_send
     200  (    1.6%)  RuboCop::Cop::Style::TrailingMethodEndStatement#on_def
     200  (    1.6%)  RuboCop::Cop::Layout::FirstArgumentIndentation#on_send
     191  (    1.5%)  RuboCop::Cop::RSpec::FactoryBot::AttributeDefinedStatically#on_block
     175  (    1.4%)  RuboCop::Cop::Layout::SpaceAroundMethodCallOperator#on_const
     164  (    1.3%)  RuboCop::Cop::Layout::IndentationConsistency#on_begin
     136  (    1.1%)  RuboCop::Cop::Layout::SpaceInsideHashLiteralBraces#on_hash
     130  (    1.0%)  RuboCop::Cop::Layout::DefEndAlignment#on_send
     113  (    0.9%)  RuboCop::Cop::Interpolation#on_dstr
     109  (    0.9%)  RuboCop::Cop::RSpec::SubjectStub#on_block
...
```

Samples number decreased from 65.3% to 56.9% and the cop samples (relative to `RuboCop::Cop::Commissioner#trigger_responding_cops` method execution) decreased from 35.7% to  0.9%.

### Total execution time

Before:
```
time bundle exec exe/rubocop --cache false --out rubocop.out .
bundle exec exe/rubocop --cache false --out rubocop.out .  83.46s user 0.40s system 99% cpu 1:24.40 total
```

After:
```
time bundle exec exe/rubocop --cache false --out rubocop.out .
bundle exec exe/rubocop --cache false --out rubocop.out .  62.60s user 0.35s system 99% cpu 1:03.21 total
```

Total execution time decreased from 1:24.40 to 1:03.21.

### Rubocop version

In performance tests was used the following Rubocop master commit:

```
ommit 7bafa099902871d9e6dc91800390a5c32e8d086b (HEAD -> master)
Author: Bozhidar Batsov <bozhidar@batsov.com>
Date:   Sat May 30 14:05:28 2020 +0300

    [Docs] Fix a couple of references to the docs folder
```

---

Before submitting the PR make sure the following are checked:

* [x] Feature branch is up-to-date with `master` (if not - rebase it).
* [ ] Squashed related commits together.
* [x] Added tests.
* [ ] Updated documentation.
* [ ] Added an entry to the [changelog](https://github.com/rubocop-hq/rubocop-rspec/blob/master/CHANGELOG.md) if the new code introduces user-observable changes.
* [x] The build (`bundle exec rake`) passes (be sure to run this locally, since it may produce updated documentation that you will need to commit).