@pirj 2020-06-07
I guess if you rebase your code, the build will pass. Fix for those has been merged https://github.com/rubocop-hq/rubocop-rspec/pull/914