@pirj 2020-06-08
Oh, sorry, I've completely missed your update in the PR description.
> Total execution time decreased from 1:24.40 to 1:03.21.

This is truly impressive 👍 
And it seems to be an incredibly slow cop.