@pirj 2020-06-08
I have never thought we'll have to optimize anytime soon, but it is happening :D
Probably it's for the best.