@andrykonchin 2020-06-08
Addressed the issue with `#on_top_leve_describe` callback.