@pirj 2020-06-08
But it only detects subject stubbing inside a `describe`, does it?
On my current project, we use a top-level `context`.