@Darhazer 2020-06-09
> But it only detects subject stubbing inside a describe, does it?
On my current project, we use a top-level context.

Maybe you got the definition of top-level describe wrong. Isn't your spec like:
```
describe Something do
  context '...' do
     ...
  end
end
```

As that is a top-level describe, and I never saw RSpec tests that begin directly with a context. If there are, other cops may not work as well