@andrykonchin 2020-06-09
> pirj
> But it only detects subject stubbing inside a describe, does it?
> On my current project, we use a top-level context.

Agree with @pirj. There are may be top-level shared examples etc... So will revert the last commit