@pirj 2020-06-09
> Maybe you got the definition of top-level describe wrong.

https://relishapp.com/rspec/rspec-core/docs/example-groups/basic-structure-describe-it surely hints that the basic structure is:
```ruby
RSpec.describe "something" do
  context "in one context" do
    it "does one thing" do
    end
  end
end
```

But how about:
```ruby
RSpec.shared_examples_for "it sabotages SubjectStub" do
  subject { haha }
  specify do
    expect(subject).to receive(:yes).and_return("no")
  end
end
```
We should detect this as well.
Is it a top-level example group? It doesn't really matter.

What about `scenario`, `fdescribe`, `xdescribe`? Why would we skip detection there?

> As that is a top-level describe, and I never saw RSpec tests that begin directly with a context.

I've never seen it before as well. But it is how it is.