@pirj 2020-06-09
My point is not to limit `SubjectStub` to just `describe`, it's not how it worked before I believe.
A quick check shows that previously `SubjectStub` was detecting offences inside top-level `xdescribe`, `fdescribe` and `feature`.

Here's the default list. https://github.com/rspec/rspec-core/blob/2b913e1e36904227e3ecc8b4af0b7c622d3cb50d/lib/rspec/core/example_group.rb#L280
https://github.com/rspec/rspec-rails/blob/a9f90932d1939502d1024db12d99f5d413091bd4/lib/rspec/rails/example/feature_example_group.rb#L49 leaving us with:
 - example_group
 - describe
 - xdescribe
 - fdescribe
 - context
 - xcontext
 - fcontext
 - feature

For some of the cops, I don't see a good reason though to leave out those https://github.com/rspec/rspec-core/blob/ddbb43674be2a02c6cc839eba15f58265abcd8ca/lib/rspec/core/shared_example_group.rb#L100:
 - shared_examples
 - shared_examples_for
 - shared_context