@pirj 2020-06-09
Please don't think that I'm defending a top-level `RSpec.context` :D I'm totally against it, as it violates https://rspec.rubystyle.guide/#context-cases at least.