@pirj 2020-06-09
> Is there any similar helper to detect any top-level example group? Not only describe?

Not that I'm aware of.

>  Does it make sense to add it?

I strongly believe so.

WDYT of:

>  or it's OK to keep manual caching

1. reverting the change involving `TopLevelDescribe`, and merging this as it was before

2. in another PR, adding a `TopLevelGroup` that would detect all, or several, depending on current usages of `TopLevelDescribe`. We'll have to keep `TopLevelDescribe` around since there can be custom cops using it. Maybe add a deprecation message?
2.1 implementing `TopLevelGroup` with performance in mind