@andrykonchin 2020-06-09
@pirj Agree with adding `TopLevelGroup` in separate PR. Will update this PR today a little bit later.