@andrykonchin 2020-06-09
Fixed all the mentioned issues. Please let me know to squash commits.