@bquorning 2020-06-11
Running `rubocop --cache=false --only RSpec/SubjectStub -- spec` to check about 1400 spec files. It took 73.19s before your change, 16.24s after. 👏