@bquorning 2020-06-11
I think this change warrants a mention in the Changelog.